Add wrapper script for korat-go
