+++ config/webpack.electron.js
@@ -4,10 +4,13 @@ const base = require('./webpack.base');
 
 const mainProcessConfig = Object.assign({}, base, {
   target: 'electron-main',
-  entry: './src/electron.ts',
+  entry: {
+    development: './src/electron.ts',
+    production: './src/electron-prod.ts',
+  },
   output: {
     path: path.join(__dirname, '../electron-built'),
-    filename: 'electron.js',
+    filename: 'electron-[name].js',
   },
 });
