+++ package.json
@@ -2,13 +2,13 @@
   "name": "korat",
   "description": "GitHub Issue Reader",
   "version": "0.1.0",
-  "main": "electron-built/electron.js",
+  "main": "electron-built/electron-production.js",
   "repository": "https://github.com/pocke/korat",
   "author": "Masataka Pocke Kuwabara <kuwabara@pocke.me>",
   "license": "Apache v2",
   "private": true,
   "scripts": {
-    "electron": "electron electron-built/electron.js",
+    "electron": "electron electron-built/electron-development.js",
     "start": "npm-run-all clean --parallel watch:main watch:renderer",
     "clean": "rm -r  public/build electron-built && mkdir -p public/build",
     "watch:renderer": "NODE_ENV=development webpack --watch --config config/webpack.renderer.js",
